Add image hashes for mintmaker, update lockfile, fix build
Supposedly if MintMaker can tell the hash of your image, it will bump it
when a newer one comes out with CVE fixes, so we know when we need to
rebuild. Freshmaker used to handle all of this for us, MintMaker seems a
lot more limited.

This adds image hashes to our dockerfile for mintmaker, updates the RPM
lockfile to match, andremoves the nonexistant package pin since we'rei
past it now

Signed-off-by: John Kyros <jkyros@redhat.com>